Use Monotorrent instead of Ragnar
replaced torrent library
updated to .net v4.6.1

+++ WebTorrentX/Models/Torrent.cs
@@ -1,21 +1,19 @@
-﻿using System;
+﻿using MonoTorrent.Client;
+using Newtonsoft.Json;
+using System;
+using System.Collections.Generic;
 using System.ComponentModel;
 using System.IO;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
-using Ragnar;
-using Newtonsoft.Json;
-using System.Collections.Generic;
-using System.Collections.ObjectModel;
 
 namespace WebTorrentX.Models
 {
     internal sealed class Torrent : INotifyPropertyChanged, IDisposable
     {
         private bool active;
-        private Session session;
-        private TorrentHandle handle;
+        private TorrentManager manager;
 
         private readonly string fastResumeDir;
         private readonly string torrentDir;
@@ -27,8 +25,8 @@ namespace WebTorrentX.Models
         {
             get
             {
-                if (handle != null && !IsEmpty)
-                    return handle.QueryStatus().Name;
+                if (manager != null && !IsEmpty)
+                    return manager.Torrent.Name;
                 else return string.Empty;
             }
         }
@@ -37,8 +35,8 @@ namespace WebTorrentX.Models
         {
             get
             {
-                if (handle != null && !IsEmpty)
-                    return Math.Round((double)handle.QueryStatus().TotalWanted / (1024 * 1024), 2);
+                if (manager != null && !IsEmpty)
+                    return Math.Round((double)manager.Torrent.Size / (1024 * 1024), 2);
                 else return 0;
             }
         }
@@ -47,9 +45,10 @@ namespace WebTorrentX.Models
         {
             get
             {
-                if (handle != null && !IsEmpty)
-                    return Math.Round((double)handle.QueryStatus().TotalWantedDone / (1024 * 1024), 2);
-                else return 0;
+                if (manager != null && !IsEmpty)
+                    return Math.Round((double)FilesInfo.Sum(x => x.DownloadedSize) / (1024 * 1024), 2);
+                else
+                    return 0;
             }
         }
 
@@ -57,8 +56,8 @@ namespace WebTorrentX.Models
         {
             get
             {
-                if (handle != null && !IsEmpty)
-                    return Math.Round(handle.QueryStatus().Progress * 100f, 2);
+                if (manager != null && !IsEmpty)
+                    return Math.Round(manager.Progress, 2);
                 else return 0;
             }
         }
@@ -67,19 +66,22 @@ namespace WebTorrentX.Models
         {
             get
             {
-                if (handle != null && !IsEmpty)
-                    return handle.QueryStatus().NumPeers;
+                if (manager != null && !IsEmpty)
+                    return manager.Torrent.HttpSeeds.Count;
                 else return 0;
             }
-        }        
+        }
+
+        MonoTorrent.SpeedMonitor speedMonitor = new MonoTorrent.SpeedMonitor();
 
         public double Speed
         {
             get
             {
-                if (handle != null && !IsEmpty)
-                    return Math.Round(handle.QueryStatus().DownloadRate / 1024f, 2);
-                else return 0;
+                if (manager != null && !IsEmpty)
+                    return Math.Round(speedMonitor.Rate / 1024f, 2);
+                else
+                    return 0;
             }
         }
 
@@ -87,7 +89,7 @@ namespace WebTorrentX.Models
         {
             get
             {
-                if (handle != null && Speed > 0 && !IsEmpty)
+                if (manager != null && Speed > 0 && !IsEmpty)
                 {
                     int p = (int)((Size - Done) * 1024 / Speed);
                     TimeSpan ts = new TimeSpan(0, 0, p);
@@ -102,76 +104,66 @@ namespace WebTorrentX.Models
         {
             get
             {
-                if (handle != null && !IsEmpty)
-                    return !handle.IsPaused;
+                if (manager != null && !IsEmpty)
+                    return manager.State != TorrentState.Paused;
                 else return false;
             }
             set
             {
-                if (handle != null && !IsEmpty)
+                if (manager != null && !IsEmpty)
                 {
-                    if (value && handle.IsPaused)
+                    if (value && manager.State == TorrentState.Paused)
                     {
-                        handle.Resume();
+                        manager.StartAsync();
                     }
-                    else if (!value && !handle.IsPaused)
+                    else if (!value && manager.State != TorrentState.Paused)
                     {
-                        handle.Pause();
-                        handle.SaveResumeData();
+                        manager.PauseAsync();
+                        //torrent.SaveResumeData();
                     }
                     UpdateProperties();
-                }                
+                }
             }
-        }   
+        }
 
         public string Status
         {
             get
             {
-                if (handle != null && !IsEmpty)
+                if (manager != null && !IsEmpty)
                 {
-                    if (handle.IsPaused)
+                    if (manager.State == TorrentState.Paused)
                         return "Pause";
                     else
-                        return handle.QueryStatus().State.ToString();
+                        return manager.State.ToString();
                 }
                 else return string.Empty;
             }
         }
 
-        public string DownloadPath
-        {
-            get
-            {
-                if (handle != null && !IsEmpty)
-                {
-                    return handle.QueryStatus().SavePath;
-                }
-                else return string.Empty;
-            }
-        }
+        public string DownloadFolder { get; private set; }
 
         public string InfoHash
         {
             get
             {
-                if (handle != null && !IsEmpty)
+                if (manager != null && !IsEmpty)
                 {
-                    return handle.InfoHash.ToHex();
+                    return manager.Torrent.InfoHash.ToHex();
                 }
                 else return string.Empty;
             }
         }
-        
+
         public IEnumerable<TorrentFileInfo> FilesInfo
         {
             get
             {
-                if (handle != null && handle.TorrentFile != null && !IsEmpty)
+                if (manager != null && !IsEmpty)
                 {
-                    for (int i = 0; i < handle.TorrentFile.NumFiles; i++)
+                    for (int i = 0; i < manager.Torrent.Files.Count(); i++)
                     {
-                        TorrentFileInfo tfinfo = new TorrentFileInfo(handle, i);
+                        TorrentFileInfo tfinfo = new TorrentFileInfo(manager, i);
                         yield return tfinfo;
                     }
                 }
@@ -185,25 +177,48 @@ namespace WebTorrentX.Models
 
         private Torrent() { }
 
-        private Torrent(AddTorrentParams addParams, Session session)
+        private Torrent(AddTorrentParams addParams)
         {
+            TorrentFileName = addParams.Filename;
+            Url = addParams.Url;
+            DownloadFolder = addParams.SaveFolder;
+
             fastResumeDir = Path.Combine(appDataFolder, ".resume");
             torrentDir = Path.Combine(appDataFolder, ".torrents");
             activeTorrentsFile = Path.Combine(appDataFolder, ".download");
             LoadTorrentState(ref addParams);
-            handle = session.AddTorrent(addParams);
-            handle.SequentialDownload = true;
-            handle.AutoManaged = false;
-            handle.FlushCache();
-            this.session = session;
-            Url = addParams.Url;
+
             active = true;
         }
 
-        public static Torrent Create(AddTorrentParams addParams, Session session)
+        public async Task StartAsync()
         {
+            var engine = new ClientEngine();
+            MonoTorrent.Torrent torrent = null;
 
-            Torrent torrent = new Torrent(addParams, session);   
+            if (File.Exists(TorrentFileName))
+            {
+                torrent = await MonoTorrent.Torrent.LoadAsync(TorrentFileName);
+            }
+            else if (!Url.StartsWith("magnet"))
+            {
+                torrent = MonoTorrent.Torrent.Load(new Uri(Url), Path.Combine(torrentDir, Path.GetTempFileName() + ".torrent"));
+            }
+
+            manager = torrent is null ? await engine.AddStreamingAsync(MonoTorrent.MagnetLink.Parse(Url), DownloadFolder)
+                 : await engine.AddStreamingAsync(torrent, DownloadFolder);
+            await manager.StartAsync();
+
+            if (!manager.HasMetadata)
+            {
+                await manager.WaitForMetadataAsync();
+            }
+        }
+
+        public static async Task<Torrent> Create(AddTorrentParams addParams)
+        {
+            Torrent torrent = new Torrent(addParams);
+            await torrent.StartAsync();
             Task.Run(delegate
             {
                 while (torrent.active)
@@ -211,15 +226,15 @@ namespace WebTorrentX.Models
                     torrent.UpdateProperties();
                     Thread.Sleep(1000);
                 }
-            });            
+            });
             return torrent;
         }
 
-        public static Torrent CreateFromSavedData(dynamic torrent, Session session)
+        public static async Task<Torrent> CreateFromSavedData(dynamic torrent)
         {
             AddTorrentParams addParams = new AddTorrentParams
             {
-                SavePath = string.IsNullOrEmpty((string)torrent.Path) ? Properties.Settings.Default.Location : (string)torrent.Path
+                SaveFolder = string.IsNullOrEmpty((string)torrent.Path) ? Properties.Settings.Default.Location : (string)torrent.Path
             };
             if (string.IsNullOrEmpty((string)torrent.TorrentFileName))
             {
@@ -235,15 +250,14 @@ namespace WebTorrentX.Models
             else
             {
                 if (File.Exists(Path.Combine(appDataFolder, ".torrents", (string)torrent.TorrentFileName)))
-                    addParams.TorrentInfo = new TorrentInfo(Path.Combine(appDataFolder, ".torrents", (string)torrent.TorrentFileName));
+                    addParams.Filename = Path.Combine(appDataFolder, ".torrents", (string)torrent.TorrentFileName);
                 else return null;
             }
-            var result =  Create(addParams, session);
-            result.Url = (string)torrent.Url;
-            result.TorrentFileName = Path.Combine(appDataFolder, ".torrents", (string)torrent.TorrentFileName);
+            var result = await Create(addParams);
+
             if (torrent.Status == "Pause")
             {
-                result.handle.Pause();
+                //result.manager.Pause();
             }
             return result;
         }
@@ -266,14 +280,14 @@ namespace WebTorrentX.Models
             OnPropertyChanged(nameof(Peers));
             OnPropertyChanged(nameof(Speed));
             OnPropertyChanged(nameof(TimeRemaining));
-            OnPropertyChanged(nameof(IsDownloading));            
+            OnPropertyChanged(nameof(IsDownloading));
             OnPropertyChanged(nameof(Status));
         }
 
         private void LoadTorrentState(ref AddTorrentParams p)
         {
-            if (IsEmpty || p.TorrentInfo == null) return;
-            string file = Path.Combine(fastResumeDir, p.TorrentInfo.Name + ".fastresume");
+            if (IsEmpty || p.Filename == null) return;
+            string file = Path.Combine(fastResumeDir, p.Filename + ".fastresume");
             if (File.Exists(file))
             {
                 p.ResumeData = File.ReadAllBytes(file);
@@ -282,30 +296,30 @@ namespace WebTorrentX.Models
 
         private void SaveTorrentState()
         {
-            if (handle.NeedSaveResumeData() && !IsEmpty)
-            {
-                handle.SaveResumeData();
-                var savedFastResume = false;
-                while (!savedFastResume)
-                {
-                    var alerts = session.Alerts.PopAll();
-                    if (alerts == null || !alerts.Any()) break;
-                    foreach (var alert in alerts)
-                    {
-                        if (alert is SaveResumeDataAlert)
-                        {
-                            var saveResumeAlert = (SaveResumeDataAlert)alert;
-                            if (!Directory.Exists(fastResumeDir)) Directory.CreateDirectory(fastResumeDir);
-                            var status = saveResumeAlert.Handle.QueryStatus();
-                            File.WriteAllBytes(
-                                Path.Combine(fastResumeDir, status.Name + ".fastresume"),
-                                saveResumeAlert.ResumeData);
-                            savedFastResume = true;
-                            break;
-                        }
-                    }
-                }
-            }            
+            //if (handle.NeedSaveResumeData() && !IsEmpty)
+            //{
+            //    handle.SaveResumeData();
+            //    var savedFastResume = false;
+            //    while (!savedFastResume)
+            //    {
+            //        var alerts = session.Alerts.PopAll();
+            //        if (alerts == null || !alerts.Any()) break;
+            //        foreach (var alert in alerts)
+            //        {
+            //            if (alert is SaveResumeDataAlert)
+            //            {
+            //                var saveResumeAlert = (SaveResumeDataAlert)alert;
+            //                if (!Directory.Exists(fastResumeDir)) Directory.CreateDirectory(fastResumeDir);
+            //                var status = saveResumeAlert.Handle.QueryStatus();
+            //                File.WriteAllBytes(
+            //                    Path.Combine(fastResumeDir, status.Name + ".fastresume"),
+            //                    saveResumeAlert.ResumeData);
+            //                savedFastResume = true;
+            //                break;
+            //            }
+            //        }
+            //    }
+            //}            
         }
 
         public event PropertyChangedEventHandler PropertyChanged;
@@ -324,7 +338,7 @@ namespace WebTorrentX.Models
                 Status = Status,
                 InfoHash = InfoHash,
                 Name = Name,
-                Path = DownloadPath,
+                Path = DownloadFolder,
                 Url = Url,
                 TorrentFileName = Path.GetFileName(TorrentFileName)
             };
@@ -335,12 +349,12 @@ namespace WebTorrentX.Models
 
         public void Remove()
         {
-            if (IsEmpty) return;
+            if (IsEmpty || manager is null) return;
             active = false;
-            string torrent = Path.Combine(torrentDir, handle.QueryStatus().Name + ".torrent");
+            string torrent = Path.Combine(torrentDir, manager.Torrent.Name + ".torrent");
             if (File.Exists(torrent))
                 File.Delete(torrent);
-            string fastResume = Path.Combine(fastResumeDir, handle.QueryStatus().Name + ".fastresume");
+            string fastResume = Path.Combine(fastResumeDir, manager.Torrent.Name + ".fastresume");
             if (File.Exists(fastResume))
                 File.Delete(fastResume);
         }
